requirements.txt
absl-py==1.2.0
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
gym==0.25.2
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
importlib-metadata==4.12.0
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
numba==0.56.2
numpy==1.21.6
oauthlib==3.2.1
pathtools==0.1.2
promise==2.3
protobuf==3.19.6
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
PyOpenGL==3.1.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.9.9
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tianshou==0.4.9.post1
torch==1.12.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.3
Werkzeug==2.2.2
zipp==3.8.1
